tests: Reorginaze unit tests
